DRV - MQTT_isReady bug fix. Removed warning.

+++ src/cmnds/cmd_if.c
@@ -154,7 +154,7 @@ float CMD_EvaluateExpression(const char *s, const char *stop) {
 	if(stop == 0) {
 		stop = s + strlen(s);
 	}
-	while(stop > s && isspace(stop[-1])) {
+	while(stop > s && isspace(((int)stop[-1]))) {
 		stop --;
 	}
 	if(g_expDebugBuffer==0){
